pyproject.toml
[project]
name = "invoiceflow-ai"
version = "0.1.0"
description = "OCR and invoice parsing project"
readme = "README.md"
authors = [
    { name="Ifunanya" }
]
requires-python = "==3.12.*"

dependencies = [
    "paddleocr",
    "paddlepaddle",
    "pillow",
    "numpy",
    "pytesseract",
    "pdfplumber",
    "pypdf",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "python-dotenv",
]

[tool.uv.extra-build-dependencies]
invoiceflow-ai = ["uv"]
